requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
djangorestframework-oauth==1.1.0
drf-social-oauth2==2.2.0
frozenlist==1.4.1
idna==3.7
jwcrypto==1.5.6
Markdown==3.6
multidict==6.0.5
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
twilio==9.0.5
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
